Fix Windows builds
